tests/test_text_to_speech.py
"""
Unit tests for text-to-speech generator
Note: These are template tests. Full implementation requires pytest and mock setup.
"""

import pytest
from pathlib import Path
import sys

# Add src to path
sys.path.insert(0, str(Path(__file__).parent.parent / 'src'))

from generators import TextToSpeechGenerator
from utils import Config


class TestTextToSpeechGenerator:
    """Test cases for TextToSpeechGenerator."""
    
    @pytest.fixture
    def generator(self):
        """Create a generator instance for testing."""
        # This would use a mock API in real tests
        return TextToSpeechGenerator()
    
    def test_initialization(self, generator):
        """Test generator initialization."""
        assert generator is not None
        assert generator.api is not None
        assert generator.converter is not None
    
    def test_generate_basic(self, generator):
        """Test basic speech generation."""
        # In real tests, mock the API call
        text = "Test speech"
        # audio = generator.generate(text)
        # assert len(audio) > 0
        pass
    
    def test_generate_with_custom_voice(self, generator):
        """Test generation with custom voice."""
        # Mock API call with specific voice
        text = "Test speech"
        voice_id = "test_voice_id"
        # audio = generator.generate(text, voice_id=voice_id)
        pass
    
    def test_batch_generation(self, generator):
        """Test batch generation."""
        texts = ["First", "Second", "Third"]
        # segments = generator.batch_generate(texts)
        # assert len(segments) == 3
        pass
    
    def test_invalid_input(self, generator):
        """Test error handling for invalid input."""
        with pytest.raises(Exception):
            # This should raise an error
            generator.generate("")
    
    def test_output_file_creation(self, generator, tmp_path):
        """Test that output files are created correctly."""
        output_file = tmp_path / "test.wav"
        # generator.generate("Test", output_file=output_file)
        # assert output_file.exists()
        pass


# To run tests:
# pytest tests/test_text_to_speech.py
