requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
elevenlabs>=1.0.0
pydub>=0.25.1
numpy>=1.24.0
